+++ src/agent_behaviours/considerations/custom_property.h
@@ -12,8 +12,8 @@ class UtilityAICustomPropertyConsideration : public UtilityAIConsiderations {
 private:
     Node*       _node_with_property;
     StringName  _property_name;
-    double      _property_max_value;    
-    double      _one_over_property_max_value;
+    float      _property_max_value;    
+    float      _one_over_property_max_value;
     Variant     _property_variant_value;
     Ref<Curve>  _activation_curve;
 protected:
@@ -32,8 +32,8 @@ public:
     void set_node_with_property( Node* node_with_property );
     Node* get_node_with_property() const;
 
-    void set_property_max_value( double true_score_value );
-    double get_property_max_value() const;
+    void set_property_max_value( float true_score_value );
+    float get_property_max_value() const;
 
     void set_property_name( StringName property_name );
     StringName get_property_name() const;
@@ -44,9 +44,9 @@ public:
    
     // Handling functions.
     void initialize_consideration();
-    virtual double evaluate() override;
-    virtual double sample_activation_curve( double input_value ) const;
-    //virtual double sample_activation_curve( double input_value ) const;
+    virtual float evaluate() override;
+    virtual float sample_activation_curve( float input_value ) const;
+    //virtual float sample_activation_curve( float input_value ) const;
 };
 
 }
