+++ src/agent_behaviours/sensor.h
@@ -23,7 +23,7 @@ public:
     
     // Handling functions. 
 
-    virtual double evaluate_sensor_value() override;
+    virtual float evaluate_sensor_value() override;
 
     // Getters and setters for attributes.
