+++ src/behaviour_tree/cooldown_usec.h
@@ -40,7 +40,7 @@ public:
 
     // Handling functions.
 
-    virtual int tick(Variant user_data, double delta) override;
+    virtual int tick(Variant user_data, float delta) override;
 };
 
 }
